Refactor chat client functionality for improved readability and maintainability
In this commit, the chat client functions were refactored in an effort to improve readability and maintainability of the code. The primary change was the separation of distinct functionality into separate methods for readability. Variable and methods names were also made more descriptive so their usage is clear. Exception handling was improved for a more robust system. An extra feature to close the emoji popup while sending was added. Moreover, a new feature for notifying the client when a connection to the server was not successful was also implemented. Implementation for displaying participant's frame was optimized. The commit also includes commenting to describe the functionality of the refactored methods.