Add message queue and improve popup implementation
In the CustomWebsocketClient class, some checks are added to make sure the mainFrame is from the MainGuiElementsInterface before proceeding. The popping mechanism has been enhanced to place messages in a queue when the message panel is not null. As a result of this change, messages won't overwrite one another and it will be easier to track all notifications in the program. The popup panel itself has been improved as well, with various functionality broken down into more manageable methods. These changes should make for a more flexible and reliable popup system.