Add custom user property and refactor emoji handling
Included a new class called CustomUserProperties. This enables chat client properties to be stored in a HashMap where the key is the sender's username. This was implemented to facilitate checking whether a message sender is already registered in the local cache.

Emoji handling was also refactored. A new method replaceEmojiDescriptionWithActualImageIcon was created in EmojiHandler class to replace the previous processText method. This significantly simplifies the process of replacing emojis in text. The method was used in various parts of the codebase where emoji processing occurs.

Finally, redundant code lines were cleaned up for readability.