Add PropertiesPanel for chat settings
A new GUI element, named 'PropertiesPanel', was added to the main gui.
The panel aims to present and allow changes to chat settings, improving user customization capability.
With this, the menu item triggering the panel was updated, and a minor update to 'ReplyPanelImpl' class was made as well.
Files associated with new GUI components were added as well.