Add nickname display functionality to chat GUI
A new feature has been implemented in GuiFunctionality.java to support the display of nicknames in the chat GUI. If a nickname is defined for a sender in the chat client properties map, it will now be displayed instead of the username. This was accomplished by introducing two new methods: checkForNickname() and displayNicknameInsteadOfUsername(). Additionally, changes were made to the determineBorderColor() method's comments for clarity.