Refactored PropertiesPanelImpl class and assigned correct username
The PropertiesPanelImpl was refactored to separate responsibilities for setting up user and client tabbed panes. The setupOwnTabbedPane and setupClientsTabbedPane methods were extracted from the constructor. This was done to improve readability and clarity of the code, as well as to allow for more effective testing of individual functionalities.

In addition, checks for null color values were added in the methods for changing border colors to prevent potential NullPointerExceptions.

Also, the hardcoded username value "own" was replaced with the actual username from the GUI in the GuiFunctionality class. This ensures that the correct username is displayed in the properties window.