Update code to support sending of images and reformatted JSON structure
The changes mainly include the support for sending images over the websocket connection. Also, the JSON structure generated for both the MessageModel and the PictureModel has been updated to have a more structured and readable format. Jackson's JsonTypeInfo and JsonSubTypes annotations were tidied. In ImagePanelImpl.java, the sendPicture() method was updated with the logic to convert an image to a byte array and send it.
A `toString()` was added to both the PictureModel and BaseModel for a more detailed output.
Lastly, the GuiFunctionality.java class was updated to print the jsonString for debugging purposes and userTextInput was modified to get the username from the GUI.