"Refactor GUI components for better readability and organization.
Changes were made in the PanelLeftImpl.java, PanelRightImpl.java and ChatMainFrameImpl.java files. They were refactored to improve readability, change the order of some methods and enhance the organization of the code. Logger variable naming was standardized for consistency. The extraction of the image from the base model was moved to a separate function for better separation of concerns. Handling of the popup menu for the editor was also grouped into their own methods to improve readability. This also holds for changes in the ChatMainFrameImpl class where some method positions were adjusted for better code organization."