Add Image processing to a separate class
Added a new file, CustomImagePanel.java, to unify image processing. This class handles extraction and resizing of images, along with functionality to open images in a separate window when clicked. This improves code modularity and simplifies classes PanelLeftImpl.java and PanelRightImpl.java, where the image processing was previously handled. This change also reflects in the CommentInterface.java which no longer needs the addMaximizePictureOnClick method.