Update application to enhance user experience and fix minor bugs
Major changes include an update in the Java and Maven versions, a more detailed guide on running the application, and some improvements in the chat functionality.
The updates in versions allow us to take advantage of the latest features and fixes. A more detailed guide on the README file is aimed at making the application easier to operate for new users.
Enhancements on the chat functionality are meant to improve application handling of message queueing and delivery, and fix a minor commit bug that caused the application to crash when the server information input by the user was invalid.