"Implement control of maximum notifications displayed"
Two new methods, getPossibleNotifications and setPossibleNotifications, along with a private variable 'possibleNotifications', were added to MainFrameInterface.java and GuiFunctionality.java classes, these methods will control the maximum number of notifications displayed at the same time on the screen.
Clean up was done in the GuiFunctionality.java by removing unnecessary line spaces and improving the readability of the code.
Redundant check for queue emptiness was removed from method checkIfDequeIsEmptyOrStartOver() in GuiFunctionality.java.
In the NotificationImpl.java class, the method to stop the timer was removed from the removeDialogAndReorderMessages() method as it's not necessary. Instead, the display of remaining notifications was added.
Minor changes included adjustments to method parameters and formatting, such as removing spaces between parameters and their values.
These changes are made to improve code readability and effectiveness of the notification system, reducing clutter on the notification pane.