Prevent empty post submission, correct field links
Implemented a check for the reply text pane to prevent users from submitting empty posts in the ReplyPanelImpl class. Added 'isTextPaneBlank' method to check if the text pane is empty before submission. This commits also fixes a bug in PanelLeftImpl and PanelRightImpl classes where timestamp was being set to the name label instead of the time label. This was a typing mistake and has been corrected to ensure correct function of panel comment essentials.