Reorder setup sections and fix time stamp handling
This commit re-orders the setup sections for `PanelLeftImpl.java`, `PanelRightImpl.java`, and respective sections in the `CustomOriginPanel.java`. The name field setup is now after the time field setup which ensures that time stamps are always setup before the name fields. This is important because this could have led to a situation where the name field got setup but the time stamp did not.

Furthermore, the handling of time stamps in `CustomOriginPanel.java` has been fixed. The time stamp was previously updated at wrong positions which could have led to inconsistent states. Now, it is only updated in the appropriate conditions, providing a more predictable and accurate time stamp management.