Add username replacement feature in custom properties
A conditional statement was added to the `CustomProperties` class to allow the replacement of a preset username called "own" only if the username instance in the main frame is not null. This change was done to support customization of user profiles while maintaining the default user in case of undefined username. The `ChatMainFrameImpl` was also altered to only override username if it's not set during startup. The function `askForUsername()` was shifted from static to dynamic in `ProgramInit` class to make it more adaptable to changes. This feature enhances user experience by making the application more flexible to user requirements.