Added method to retrieve version from JAR
I have added a new method 'retrieveJarVersion' which can get the version number from the properties file inside a JAR. I also updated the 'repaintMainFrame' method to include the version number in the frame title.
This will help in displaying the correct version of the application in the window and managing version control more effectively. In addition, I have rearranged some methods to improve code readability.