Update JFormDesigner and Java versions and remove checkbox state
The versions of JFormDesigner and Java were downgraded to "8.0.3.0.246" and "17.0.3" respectively. This was due to compatibility issues discovered with the newer versions in our current environment.

Additionally, the 'selected' state of 'externalNotificationsMenuItem' checkbox is removed. This was causing unintended behavior in the user interface where the checkbox was always selected by default, disregarding the user preferences.