Refactor notification handling in GUI functionality
This commit refactors how the messages and notifications are handled in the GuiFunctionality class. Changes include the addition of new sections for adding messages to the queue, writing to a pane, and creating desktop notifications. These changes help to streamline functionality and improve user experience by ensuring messages and notifications are processed correctly.

Concerns to the systems are also reflected, specifically, the placement of TODO to point out that currently, the external notification handling functionality only supports Linux and needs Windows support.