Add CommentManager interface, refactor classes and improve message handling
Implemented a CommentManager interface to abstract the comment handling for improving code readability and maintainability. New services are refactored to use this interface. Adjusted the GuiFunctionality, ProgramInit and CustomWebsocketClient to handle changes. Also, introduced MessageCategory for message organization.
Improved the processAndDisplayMessage method to ease the logic flow. These changes aim to make the code cleaner and easier to manage.