"Refactored code for Comments' implementation
This commit enhances the code readability and maintainability by refactoring part of the code sections related to the comments' implementation. Code sections related to Text, Link and Picture have been separated in the setup methods. Moreover, a new class, namely LinkLeftImpl, has been added to represent the concept of a link message to the left. In addition to these, some excess methods have been removed and minor UI adjustments have been done as well for a cohesive user interface. Lastly, some changes were required in the processing of a message coming from a socket."