Add link support and dependency in chat system
Added support for detecting and handling HTTP/HTTPS links in the chat system. This feature includes the creation of a confirmation dialog when a link is detected. This feature will improve user experience by providing a clickable link and prevents the accidental or unintended sending of links. Furthermore, the Jsoup dependency was included to fetch and display metadata of linked web pages. Changes were made in various classes such as GuiFunctionalityImpl.java, pom.xml, and additional dialog interfaces were added. Also, the version has been updated from v1.0.2 to v.1.0.3 in pom.xml and README.md. The method `testIfSystemDesktopEnvironmentIsDetected` from MainFrameTests.java was removed since it was redundant and did not contribute to the functionality.