Add emoji replacement function and clean up tests - bug fix
This commit adds a new functionality to the chat application, specifically in the ChatMainFrameImpl class, that will replace any emoji icons with their text descriptions. This is to provide a more accessible system for all users.

In addition, several unnecessary and redundant tests in the MainFrameTests class were removed to streamline the testing process and increase readability. This will lead to easier code maintenance and less confusion going forward. In order to maintain full testing coverage, new tests would need to be created to address the new emoji replacement functionality.