Refactor GuiFunctionalityImpl for readability, fix 'send' functionality
This commit refactors the GuiFunctionalityImpl class for better readability and maintainability. Unnecessary comments and unused code were cleaned, and several methods were simplified or broken down to smaller methods. Emphasizing code quality, this commit significantly improves readability without changing the core functionality.

Also, a noticeable change was made in the 'send' functionality in both the ChatMainFrameImpl and GuiFunctionalityImpl classes. The method clearTextPaneAndSendMessageToSocket was separated into two distinct methods; sendMessageToSocket and clearTextPane, respectively. This ensures single responsibility per method and improves code clarity. The change was necessary as it was