Refactor classes and methods; Add interrupt feature
Various classes and methods across modules were significantly refactored for better readability, maintainability, and adherence to coding standards. Import paths in `CommentManagerImpl` and `GuiFunctionalityImpl` were adjusted due to the restructuring of classes and methods. Detailed JavaDoc was added to methods in `GuiFunctionalityImpl` for clarity. The `ByteArrayHandler` class was introduced to manage byte array operations.

Additionally, an 'interrupt' feature was implemented. This includes the addition of a menu item in the main GUI along with corresponding handling in the backend. This feature intends to allow users to interrupt or get the attention of other users in the application. Due to the volume of changes, careful review is advised before merge.