Add InterruptHandler and refactor existing code
Implemented a new feature for interrupting chat, by introducing `InterruptHandler` and `InterrupHandlerInterface`. The goal was to decouple the code related to interruption from the `GuiFunctionalityImpl` class. Specifically, the logic to extract usernames from JSON and force the chat GUI to be in the front has been moved to the `InterruptHandler`.

Simultaneously, several code lines in `GuiFunctionalityImpl` have been refactored to associate with the newly added classes. This not only improves readability but also adheres to the Single Responsibility Principle, paving the way for future enhancements in this area.