Add CacheManager setter in MessageDisplayHandlerInterface and tests
MessageDisplayHandlerInterface has been extended with a method
to explicitly set the CacheManager.
Previously,
the CacheManager was initialized directly in the constructor of the implementing classes via the CacheManagerFactory.
This change decouples the code from the factory
and improves testability
by allowing the injection of mock CacheManager implementations for testing purposes.
This commitment also adds tests for the new method
and uses it in GuiFunctionalityImpl.