Refactor CustomProperties to Singleton Pattern
Refactored the CustomProperties class to the Singleton pattern to ensure only one instance of CustomProperties is used throughout the application. This change was crucial to maintain consistency across all parts of the application using this class. Some unnecessary imports were also removed for code cleanliness. Moreover, modified the way in which custom properties of the chat client are retrieved and saved. Instead of calling these methods on the MainFrameInterface, they are now called directly on the Singleton instance. Lastly, added a "propertiescache" case to CacheManager to persist properties data.