Improve user properties initialization and loading process
This change refactors the initialization and loading process of user properties. The import statements have been adjusted to support the new 'CustomProperties' class instead of the 'CustomPropertiesInterface'.

The main changes occur in the 'ProgramInit' class, in which the 'initializeProperties' method has been extended with 'envVariables' as an additional parameter. The existence of a configuration file is now checked before property loading and an instance of 'CustomUserProperties' is added for the current user.

The 'initializeMainFrame' method has also been slightly rearranged to improve the sequence of initialization steps. Now, it allowed for user name loading to occur after main frame initialization, rather than before.

Further modifications include the change of return type of 'checkIfConfigFileExists' method and some reordering of methods in the 'ChatMainFrameImpl' class for better readability.

This change addresses previous ambiguities in the properties initialization process and makes it more efficient. It also makes the code cleaner and easier to understand.