Refactor CustomUserPropertiesDTO and fix typos
The fundamental change was the refactor of the CustomUserPropertiesDTO. Introduced fluent methods to replace values of DTO object which eliminates the need to create new instances each time. This promotes cleaner and more readable code. The methods added were: 'withUsername()', 'withNickname()', and 'withBorderColor()'. Also, a typo in ProgramInit.java file where 'initiliazeMainFrame()' was changed to 'initializeMainFrame()'.