Add new LinkTransfer Data Transfer Object
Added a new LinkTransferDTO object to replace the LinkCommentRecord object for transferring links and comments between the client and the web socket server. This change enhances code readability and maintainability by providing a more descriptive and appropriate name for the object. The LinkTransferDTO is used in the LinkDialogImpl, LinkRightImpl, and LinkLeftImpl classes where links and comments are sent and received. A JSON processing exception is now also caught and handled when deserializing string messages back into LinkTransferDTO objects.

The code was also cleaned up to remove unnecessary comments and improve readability.

Re-positioning of the chat frame for testing purposes was also improved for readability, and work was done on the initialization of emoji functionality.

This commit continues efforts of improving code quality and readability.