Refactor comment system for improved user interactions
Order of operations was adjusted in the CommentManagerImpl class to avoid null pointer exceptions. A method (updateExistingMessage) was added to the MessageDisplayHandlerInterface to handle updated message functionality. CacheManager implementation was refactored to handle normal and interacted messages accordingly in GuiFunctionalityImpl. PanelRightImpl class has now methods to set and get BaseModel for further manipulation of comments, as does PanelLeftImpl. ReactionStickerImpl has a method to adjust visibility of stickers based on available slots. Panel components in MessageDisplayHandler can now be updated with new BaseModel references. This broad refactor improves the handling of user interactions within the comment system, specifically ensuring the correct display and updates of reaction stickers on user messages.