Add CustomTransferHandler and refactor related classes
Introduced a new CustomTransferHandler that handles dropped data onto the text editor pane of the GUI. This significantly simplified the GuiFunctionalityImpl class as the logic for distinguishing between image, normal text data as well as links and triggering corresponding actions has been moved to the new CustomTransferHandler.

Major adjustments have been made in ChatMainFrameImpl and GuiFunctionalityImpl to delegate related responsibilities to the new class. ChatMainFrameImpl now only sends messages to the socket instead of triggering different actions. This change has made the code easier to maintain by encapsulating related actions within one class.

GuiFunctionalityInterfaceTest has been added to validate the text retrieval from the input pane. This provides confidence in the reliability of the getTextFromInput method in GuiFunctionalityImpl, which is now solely responsible for obtaining input from the user.

The GuiFunctionalityInterface got updated. Methods that do not need to be exposed anymore were removed, improving encapsulation. The 'sendMessageToSocket' method now takes a parameter to convey a clear responsibility.

This commit improves the overall design by adhering to the Single Responsibility Principle and makes the code cleaner and easier to maintain.